# swagger-api-2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "shop-food-api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/wh/supplier/create": {
      "post": {
        "tags": [
          "SupplierWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierWhCreateModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierWhCreateModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierWhCreateModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierWhCreateModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierWhCreateModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierWhCreateModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/supplier/update": {
      "post": {
        "tags": [
          "SupplierWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierWhUpdateModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierWhUpdateModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierWhUpdateModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierWhUpdateModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierWhUpdateModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierWhUpdateModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/supplier/delete": {
      "post": {
        "tags": [
          "SupplierWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierWhDeleteModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierWhDeleteModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierWhDeleteModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierWhDeleteModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierWhDeleteModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierWhDeleteModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/supplier/list": {
      "post": {
        "tags": [
          "SupplierWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierWhListModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierWhListModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierWhListModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierWhListModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierWhListModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierWhListModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/transaction/create": {
      "post": {
        "tags": [
          "TransactionWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionWhCreateModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionWhCreateModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionWhCreateModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionWhCreateModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionWhCreateModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionWhCreateModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/transaction/update": {
      "post": {
        "tags": [
          "TransactionWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionWhUpdateModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionWhUpdateModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionWhUpdateModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionWhUpdateModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionWhUpdateModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionWhUpdateModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/transaction/delete": {
      "post": {
        "tags": [
          "TransactionWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionWhDeleteModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionWhDeleteModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionWhDeleteModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionWhDeleteModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionWhDeleteModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionWhDeleteModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/transaction/list": {
      "post": {
        "tags": [
          "TransactionWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionWhListModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionWhListModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionWhListModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionWhListModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionWhListModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionWhListModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/wh/warehouse/create": {
      "post": {
        "tags": [
          "WarehouseWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseCreateModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseCreateModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseCreateModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseCreateModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseCreateModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseCreateModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/warehouse/update": {
      "post": {
        "tags": [
          "WarehouseWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseUpdateModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseUpdateModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseUpdateModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseUpdateModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseUpdateModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseUpdateModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/warehouse/delete": {
      "post": {
        "tags": [
          "WarehouseWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseDeleteModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseDeleteModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseDeleteModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseDeleteModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseDeleteModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseDeleteModelResApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/wh/warehouse/list": {
      "post": {
        "tags": [
          "WarehouseWh"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseListModelReq"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseListModelReq"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WarehouseListModelReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseListModelResApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseListModelResApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarehouseListModelResApiResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "MetaData": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "exceptionExtra": {
            "$ref": "#/components/schemas/Exception"
          }
        },
        "additionalProperties": false
      },
...
      "SupplierWhCreateModelReq": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SupplierWhCreateModelRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdBy": {
            "type": "string",
            "nullable": true
          },
          "updatedBy": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SupplierWhCreateModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SupplierWhCreateModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "SupplierWhDeleteModelReq": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "SupplierWhDeleteModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "SupplierWhDeleteModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SupplierWhDeleteModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "SupplierWhListModelReq": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "SupplierWhListModelRes": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SupplierModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SupplierWhListModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SupplierWhListModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "SupplierWhUpdateModelReq": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SupplierWhUpdateModelRes": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SupplierWhUpdateModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SupplierWhUpdateModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "TransactionWhCreateModelReq": {
        "type": "object",
        "additionalProperties": false
      },
      "TransactionWhCreateModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "TransactionWhCreateModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TransactionWhCreateModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "TransactionWhDeleteModelReq": {
        "type": "object",
        "additionalProperties": false
      },
      "TransactionWhDeleteModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "TransactionWhDeleteModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TransactionWhDeleteModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "TransactionWhListModelReq": {
        "type": "object",
        "additionalProperties": false
      },
      "TransactionWhListModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "TransactionWhListModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TransactionWhListModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "TransactionWhUpdateModelReq": {
        "type": "object",
        "additionalProperties": false
      },
      "TransactionWhUpdateModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "TransactionWhUpdateModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TransactionWhUpdateModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
...
      "WarehouseCreateModelReq": {
        "required": [
          "address",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "address": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "WarehouseCreateModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "WarehouseCreateModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WarehouseCreateModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "WarehouseDeleteModelReq": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "WarehouseDeleteModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "WarehouseDeleteModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WarehouseDeleteModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "WarehouseInfoModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WarehouseListModelReq": {
        "type": "object",
        "properties": {
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "WarehouseListModelRes": {
        "type": "object",
        "properties": {
          "listWarehouse": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WarehouseInfoModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WarehouseListModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WarehouseListModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      },
      "WarehouseUpdateModelReq": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WarehouseUpdateModelRes": {
        "type": "object",
        "additionalProperties": false
      },
      "WarehouseUpdateModelResApiResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WarehouseUpdateModelRes"
          },
          "isNormal": {
            "type": "boolean"
          },
          "metaData": {
            "$ref": "#/components/schemas/MetaData"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}